* refact: Alter MessageHolder to correspond to Notification pattern
- Change MessageHolder ErrorCode to int: As notification pattern it's about domain logic, keep error code as HttpStatusCode is a SOLID offense and also a "Ports and Adapters" offense.
SOLID because, violates the Single Responsibility Principle, keeping notifications and resolving http status returns.
Ports and Adapters because the core layer (generally, the domain layer or "Business in our case") has to presume what is the driver adapter (HTTP in this case).

- Change Notification from Shared to Business layer: Despistes driven layers have their it specific errors and exceptions, they should be translated to Domain errors.
In this way, domain can handle it and/or by passing it.
Considering that, make sense the notification mechanism stay in Business layer and ErrorCode specification be there too.
Thus, every driver or driven layer can know what error domain can handle.

- Map error codes to HttpStatusCodes: API layers should map any ErroCodes

- Rename MessageHolder to Notification: This denomination seems to be more appropriate.